combine builder methods in bind method
