+++ buf.gen.yaml
@@ -1,7 +1,13 @@
 version: v1
 plugins:
-  - name: js
-    out: gen/js
-    opt: import_style=es6
-  - name: ts
-    out: gen/ts
+  # - name: js
+  #   out: gen/js
+  #   opt: import_style=es6
+  # - name: ts
+  #   out: gen/ts
+  - remote: buf.build/cfacorp/plugins/swift-protobuf:v1.19.0-1
+    out: gen/swift
+    opt:
+      - Visibility=Public
+  - name: kotlin
+    out: gen/kotlin
