+++ buf.work.yaml
@@ -1,4 +1,3 @@
 version: v1
 directories:
   - proto
-  - pkg/motor/proto
