+++ proto/motor/v1/request.proto
@@ -7,6 +7,9 @@ import "schema/v1/schema.proto";
 
 option go_package = "github.com/sonr-io/sonr/pkg/motor/types";
 
+option java_package = "io.sonr.types";
+option java_multiple_files = true;
+
 // -----------------------------------------------------------------------------
 // Motor Node API
 // -----------------------------------------------------------------------------
